+++ src/main/java/com/openggf/game/sonic3k/events/S3kTransitionEventBridge.java
@@ -30,5 +30,5 @@ public interface S3kTransitionEventBridge {
 
     void requestMgzPostTransitionRelease();
 
-    void requestCnzPostTransitionRelease(int framesUntilRelease);
+    void requestCnzPostTransitionRelease();
 }